feat: Add react-icons dependency for icon component support (#123)
Add `react-icons` package as a new dependency in `package.json` and
`package-lock.json`. This allows for the usage of pre-built icons in the
application's UI components. The new dependency is at version 4.12.0 and
includes a peer dependency of React.

In addition, the `src/components/nav/nav-list.tsx` file has been
modified to import specific icons from `react-icons` and use them in the
corresponding UI elements for visual enhancements.

This change improves the user experience by providing a wider range of
customizable icons, ultimately enhancing the visual appeal of the
application's interface.

Fixes #123