feat: Add FAQ field to AIConfig model and remove FAQ Bot from nav
The FAQ field is added to the AIConfig model in order to store FAQ data
associated with a particular user. This allows for easier management and
retrieval of FAQ information. Additionally, the FAQ Bot section is
removed from the navigation list as it is no longer necessary. These
changes improve the organization and functionality of the application.

Fixes #123