Refactor popoverFormFaq component UI for response input and file attachment
This change updates the UI of the popoverFormFaq component for handling
response input and file attachment. Previously, the response input field
and the option to attach a file were displayed separately. Now, a
checkbox has been added to toggle between them. When the checkbox is
unchecked, the user can input the response directly into a text field.
When it is checked, the user can select a file, which will be read as a
data URL and assigned to the response. This refactor improves the user
experience by providing a more streamlined and intuitive interface for
inputting responses.

No issue references are associated with this change.